chore: move trust subcommand to folder
