+++ lib/commands/trust/index.js
@@ -1,14 +1,14 @@
-const BaseCommand = require('../base-cmd.js')
+const BaseCommand = require('../../base-cmd.js')
 
 class Trust extends BaseCommand {
   static description = 'Create a trusted relationship between a package and a OIDC provider'
   static name = 'trust'
 
   static subcommands = {
-    github: require('../subcommands/trust-github.js'),
-    gitlab: require('../subcommands/trust-gitlab.js'),
-    list: require('../subcommands/trust-list.js'),
-    revoke: require('../subcommands/trust-revoke.js'),
+    github: require('./trust-github.js'),
+    gitlab: require('./trust-gitlab.js'),
+    list: require('./trust-list.js'),
+    revoke: require('./trust-revoke.js'),
   }
 
   static async completion (opts) {
