+++ lib/commands/trust/trust-github.js
@@ -1,22 +1,22 @@
 const Definition = require('@npmcli/config/lib/definitions/definition.js')
 const globalDefinitions = require('@npmcli/config/lib/definitions/definitions.js')
-const TrustCommand = require('../trust-cmd.js')
+const TrustCommand = require('../../trust-cmd.js')
 const path = require('node:path')
 
-class TrustGitLab extends TrustCommand {
-  static description = 'Create a trusted relationship between a package and GitLab CI/CD'
-  static name = 'gitlab'
+class TrustGitHub extends TrustCommand {
+  static description = 'Create a trusted relationship between a package and GitHub Actions'
+  static name = 'github'
   static positionals = 1 // expects at most 1 positional (package name)
-  static providerName = 'GitLab CI/CD'
-  static providerEntity = 'GitLab project'
-  static providerFile = 'GitLab CI/CD Pipeline'
-  static providerHostname = 'https://gitlab.com'
+  static providerName = 'GitHub Actions'
+  static providerEntity = 'GitHub repository'
+  static providerFile = 'GitHub Actions Workflow'
+  static providerHostname = 'https://github.com'
 
   // entity means project / repository
-  static entityKey = 'project'
+  static entityKey = 'repository'
 
   static usage = [
-    '[package] --file [--project|--repo|--repository] [--env|--environment] [-y|--yes]',
+    '[package] --file [--repo|--repository] [--env|--environment] [-y|--yes]',
   ]
 
   static definitions = {
@@ -26,12 +26,13 @@ class TrustGitLab extends TrustCommand {
     file: new Definition('file', {
       default: null,
       type: String,
-      description: 'Name of pipeline file (e.g., .gitlab-ci.yml)',
+      description: 'Name of workflow file within a repositories .GitHub folder (must end in yaml, yml)',
     }),
-    project: new Definition('project', {
+    repository: new Definition('repository', {
       default: null,
       type: String,
-      description: 'Name of the project in the format group/project or group/subgroup/project',
+      description: 'Name of the repository in the format owner/repo',
+      alias: ['repo'],
     }),
     environment: new Definition('environment', {
       default: null,
@@ -43,32 +44,30 @@ class TrustGitLab extends TrustCommand {
 
   getEntityUrl ({ providerHostname, file, entity }) {
     if (file) {
-      return new URL(`${entity}/-/blob/HEAD/${file}`, providerHostname).toString()
+      return new URL(`${entity}/blob/HEAD/.github/workflows/${file}`, providerHostname).toString()
     }
     return new URL(entity, providerHostname).toString()
   }
 
   validateEntity (entity) {
-    if (entity.split('/').length < 2) {
-      throw new Error(`${this.providerEntity} must be specified in the format group/project or group/subgroup/project`)
+    if (entity.split('/').length !== 2) {
+      throw new Error(`${this.providerEntity} must be specified in the format owner/repository`)
     }
   }
 
   validateFile (file) {
     if (file !== path.basename(file)) {
-      throw new Error('GitLab CI/CD pipeline file must be just a file not a path')
+      throw new Error('GitHub Actions workflow must be just a file not a path')
     }
   }
 
   static optionsToBody (options) {
-    const { file, project, environment } = options
+    const { file, repository, environment } = options
     const trustConfig = {
-      type: 'gitlab',
+      type: 'github',
       claims: {
-        project_path: project,
-        // this looks off, but this is correct
-        /** The ref path to the top-level pipeline definition, for example, gitlab.example.com/my-group/my-project//.gitlab-ci.yml@refs/heads/main. Introduced in GitLab 16.2. This claim is null unless the pipeline definition is located in the same project. */
-        ci_config_ref_uri: {
+        repository,
+        workflow_ref: {
           file,
         },
       },
@@ -81,14 +80,14 @@ class TrustGitLab extends TrustCommand {
 
   // Convert API response body to options
   static bodyToOptions (body) {
-    const file = body.claims?.ci_config_ref_uri?.file
-    const project = body.claims?.project_path
+    const file = body.claims?.workflow_ref?.file
+    const repository = body.claims?.repository
     const environment = body.environment
     return {
       ...(body.id) && { id: body.id },
       ...(body.type) && { type: body.type },
       ...(file) && { file },
-      ...(project) && { project },
+      ...(repository) && { repository },
       ...(environment) && { environment },
     }
   }
@@ -101,4 +100,4 @@ class TrustGitLab extends TrustCommand {
   }
 }
 
-module.exports = TrustGitLab
+module.exports = TrustGitHub
