+++ lib/commands/trust/trust-gitlab.js
@@ -1,22 +1,22 @@
 const Definition = require('@npmcli/config/lib/definitions/definition.js')
 const globalDefinitions = require('@npmcli/config/lib/definitions/definitions.js')
-const TrustCommand = require('../trust-cmd.js')
+const TrustCommand = require('../../trust-cmd.js')
 const path = require('node:path')
 
-class TrustGitHub extends TrustCommand {
-  static description = 'Create a trusted relationship between a package and GitHub Actions'
-  static name = 'github'
+class TrustGitLab extends TrustCommand {
+  static description = 'Create a trusted relationship between a package and GitLab CI/CD'
+  static name = 'gitlab'
   static positionals = 1 // expects at most 1 positional (package name)
-  static providerName = 'GitHub Actions'
-  static providerEntity = 'GitHub repository'
-  static providerFile = 'GitHub Actions Workflow'
-  static providerHostname = 'https://github.com'
+  static providerName = 'GitLab CI/CD'
+  static providerEntity = 'GitLab project'
+  static providerFile = 'GitLab CI/CD Pipeline'
+  static providerHostname = 'https://gitlab.com'
 
   // entity means project / repository
-  static entityKey = 'repository'
+  static entityKey = 'project'
 
   static usage = [
-    '[package] --file [--repo|--repository] [--env|--environment] [-y|--yes]',
+    '[package] --file [--project|--repo|--repository] [--env|--environment] [-y|--yes]',
   ]
 
   static definitions = {
@@ -26,13 +26,12 @@ class TrustGitHub extends TrustCommand {
     file: new Definition('file', {
       default: null,
       type: String,
-      description: 'Name of workflow file within a repositories .GitHub folder (must end in yaml, yml)',
+      description: 'Name of pipeline file (e.g., .gitlab-ci.yml)',
     }),
-    repository: new Definition('repository', {
+    project: new Definition('project', {
       default: null,
       type: String,
-      description: 'Name of the repository in the format owner/repo',
-      alias: ['repo'],
+      description: 'Name of the project in the format group/project or group/subgroup/project',
     }),
     environment: new Definition('environment', {
       default: null,
@@ -44,30 +43,32 @@ class TrustGitHub extends TrustCommand {
 
   getEntityUrl ({ providerHostname, file, entity }) {
     if (file) {
-      return new URL(`${entity}/blob/HEAD/.github/workflows/${file}`, providerHostname).toString()
+      return new URL(`${entity}/-/blob/HEAD/${file}`, providerHostname).toString()
     }
     return new URL(entity, providerHostname).toString()
   }
 
   validateEntity (entity) {
-    if (entity.split('/').length !== 2) {
-      throw new Error(`${this.providerEntity} must be specified in the format owner/repository`)
+    if (entity.split('/').length < 2) {
+      throw new Error(`${this.providerEntity} must be specified in the format group/project or group/subgroup/project`)
     }
   }
 
   validateFile (file) {
     if (file !== path.basename(file)) {
-      throw new Error('GitHub Actions workflow must be just a file not a path')
+      throw new Error('GitLab CI/CD pipeline file must be just a file not a path')
     }
   }
 
   static optionsToBody (options) {
-    const { file, repository, environment } = options
+    const { file, project, environment } = options
     const trustConfig = {
-      type: 'github',
+      type: 'gitlab',
       claims: {
-        repository,
-        workflow_ref: {
+        project_path: project,
+        // this looks off, but this is correct
+        /** The ref path to the top-level pipeline definition, for example, gitlab.example.com/my-group/my-project//.gitlab-ci.yml@refs/heads/main. Introduced in GitLab 16.2. This claim is null unless the pipeline definition is located in the same project. */
+        ci_config_ref_uri: {
           file,
         },
       },
@@ -80,14 +81,14 @@ class TrustGitHub extends TrustCommand {
 
   // Convert API response body to options
   static bodyToOptions (body) {
-    const file = body.claims?.workflow_ref?.file
-    const repository = body.claims?.repository
+    const file = body.claims?.ci_config_ref_uri?.file
+    const project = body.claims?.project_path
     const environment = body.environment
     return {
       ...(body.id) && { id: body.id },
       ...(body.type) && { type: body.type },
       ...(file) && { file },
-      ...(repository) && { repository },
+      ...(project) && { project },
       ...(environment) && { environment },
     }
   }
@@ -100,4 +101,4 @@ class TrustGitHub extends TrustCommand {
   }
 }
 
-module.exports = TrustGitHub
+module.exports = TrustGitLab
