+++ lib/commands/trust/trust-list.js
@@ -1,10 +1,10 @@
 const pkgJson = require('@npmcli/package-json')
-const { otplease } = require('../utils/auth.js')
+const { otplease } = require('../../utils/auth.js')
 const npmFetch = require('npm-registry-fetch')
 const npa = require('npm-package-arg')
 const TrustGithub = require('./trust-github.js')
 const TrustGitlab = require('./trust-gitlab.js')
-const TrustCommand = require('../trust-cmd.js')
+const TrustCommand = require('../../trust-cmd.js')
 const globalDefinitions = require('@npmcli/config/lib/definitions/definitions.js')
 
 // Convert trust config body to options object
