+++ lib/commands/trust/trust-revoke.js
@@ -1,9 +1,9 @@
 const globalDefinitions = require('@npmcli/config/lib/definitions/definitions.js')
 const Definition = require('@npmcli/config/lib/definitions/definition.js')
 const pkgJson = require('@npmcli/package-json')
-const { otplease } = require('../utils/auth.js')
+const { otplease } = require('../../utils/auth.js')
 const npmFetch = require('npm-registry-fetch')
-const BaseCommand = require('../base-cmd.js')
+const BaseCommand = require('../../base-cmd.js')
 const npa = require('npm-package-arg')
 const { log } = require('proc-log')
