+++ lib/npm.js
@@ -26,7 +26,7 @@ class Npm {
         command: c,
       })
     }
-    return require(`./commands/${command}.js`)
+    return require(`./commands/${command}`)
   }
 
   unrefPromises = []
