+++ test/lib/commands/trust/github.js
@@ -1,11 +1,11 @@
 const t = require('tap')
-const { load: loadMockNpm } = require('../../fixtures/mock-npm.js')
+const { load: loadMockNpm } = require('../../../fixtures/mock-npm.js')
 const MockRegistry = require('@npmcli/mock-registry')
 const realProcLog = require('proc-log')
 
 const packageName = '@npmcli/test-package'
 
-t.test('trust-gitlab with all options provided', async t => {
+t.test('github with all options provided', async t => {
   const { npm } = await loadMockNpm(t, {
     prefixDir: {
       'package.json': JSON.stringify({
@@ -35,10 +35,10 @@ t.test('trust-gitlab with all options provided', async t => {
 
   registry.trustCreate({ packageName })
 
-  await npm.exec('trust', ['gitlab', packageName, '--yes', '--file', '.gitlab-ci.yml', '--project', 'group/subgroup/repo', '--environment', 'production'])
+  await npm.exec('trust', ['github', packageName, '--yes', '--file', 'workflow.yml', '--repository', 'owner/repo', '--environment', 'production'])
 })
 
-t.test('trust-gitlab with invalid project format', async t => {
+t.test('github with invalid repository format', async t => {
   const { npm } = await loadMockNpm(t, {
     prefixDir: {
       'package.json': JSON.stringify({
@@ -61,12 +61,12 @@ t.test('trust-gitlab with invalid project format', async t => {
   })
 
   await t.rejects(
-    npm.exec('trust', ['gitlab', packageName, '--yes', '--file', '.gitlab-ci.yml', '--project', 'invalid']),
-    { message: /must be specified in the format/ }
+    npm.exec('trust', ['github', packageName, '--yes', '--file', 'workflow.yml', '--repository', 'invalid']),
+    { message: /must be specified in the format owner\/repository/ }
   )
 })
 
-t.test('trust-gitlab with file as path', async t => {
+t.test('github with file as path', async t => {
   const { npm } = await loadMockNpm(t, {
     prefixDir: {
       'package.json': JSON.stringify({
@@ -89,12 +89,12 @@ t.test('trust-gitlab with file as path', async t => {
   })
 
   await t.rejects(
-    npm.exec('trust', ['gitlab', packageName, '--yes', '--file', '.gitlab/ci.yml', '--project', 'group/repo']),
+    npm.exec('trust', ['github', packageName, '--yes', '--file', '.github/workflows/ci.yml', '--repository', 'owner/repo']),
     { message: /must be just a file not a path/ }
   )
 })
 
-t.test('trust-gitlab without environment', async t => {
+t.test('github without environment', async t => {
   const { npm } = await loadMockNpm(t, {
     prefixDir: {
       'package.json': JSON.stringify({
@@ -124,30 +124,30 @@ t.test('trust-gitlab without environment', async t => {
 
   registry.trustCreate({ packageName })
 
-  await npm.exec('trust', ['gitlab', packageName, '--yes', '--file', '.gitlab-ci.yml', '--project', 'group/repo'])
+  await npm.exec('trust', ['github', packageName, '--yes', '--file', 'workflow.yml', '--repository', 'owner/repo'])
 })
 
 t.test('bodyToOptions with all fields', t => {
-  const TrustGitLab = require('../../../lib/subcommands/trust-gitlab.js')
+  const TrustGitHub = require('../../../../lib/commands/trust/trust-github.js')
 
   const body = {
     id: 'test-id',
-    type: 'gitlab',
+    type: 'github',
     claims: {
-      project_path: 'group/repo',
-      ci_config_ref_uri: {
-        file: '.gitlab-ci.yml',
+      repository: 'owner/repo',
+      workflow_ref: {
+        file: 'test.yml',
       },
     },
     environment: 'prod',
   }
 
-  const options = TrustGitLab.bodyToOptions(body)
+  const options = TrustGitHub.bodyToOptions(body)
 
   t.equal(options.id, 'test-id', 'id should be set')
-  t.equal(options.type, 'gitlab', 'type should be set')
-  t.equal(options.file, '.gitlab-ci.yml', 'file should be set')
-  t.equal(options.project, 'group/repo', 'project should be set')
+  t.equal(options.type, 'github', 'type should be set')
+  t.equal(options.file, 'test.yml', 'file should be set')
+  t.equal(options.repository, 'owner/repo', 'repository should be set')
   t.equal(options.environment, 'prod', 'environment should be set')
   t.end()
 })
