+++ test/lib/commands/trust/gitlab.js
@@ -1,11 +1,11 @@
 const t = require('tap')
-const { load: loadMockNpm } = require('../../fixtures/mock-npm.js')
+const { load: loadMockNpm } = require('../../../fixtures/mock-npm.js')
 const MockRegistry = require('@npmcli/mock-registry')
 const realProcLog = require('proc-log')
 
 const packageName = '@npmcli/test-package'
 
-t.test('trust-github with all options provided', async t => {
+t.test('gitlab with all options provided', async t => {
   const { npm } = await loadMockNpm(t, {
     prefixDir: {
       'package.json': JSON.stringify({
@@ -35,10 +35,10 @@ t.test('trust-github with all options provided', async t => {
 
   registry.trustCreate({ packageName })
 
-  await npm.exec('trust', ['github', packageName, '--yes', '--file', 'workflow.yml', '--repository', 'owner/repo', '--environment', 'production'])
+  await npm.exec('trust', ['gitlab', packageName, '--yes', '--file', '.gitlab-ci.yml', '--project', 'group/subgroup/repo', '--environment', 'production'])
 })
 
-t.test('trust-github with invalid repository format', async t => {
+t.test('gitlab with invalid project format', async t => {
   const { npm } = await loadMockNpm(t, {
     prefixDir: {
       'package.json': JSON.stringify({
@@ -61,12 +61,12 @@ t.test('trust-github with invalid repository format', async t => {
   })
 
   await t.rejects(
-    npm.exec('trust', ['github', packageName, '--yes', '--file', 'workflow.yml', '--repository', 'invalid']),
-    { message: /must be specified in the format owner\/repository/ }
+    npm.exec('trust', ['gitlab', packageName, '--yes', '--file', '.gitlab-ci.yml', '--project', 'invalid']),
+    { message: /must be specified in the format/ }
   )
 })
 
-t.test('trust-github with file as path', async t => {
+t.test('gitlab with file as path', async t => {
   const { npm } = await loadMockNpm(t, {
     prefixDir: {
       'package.json': JSON.stringify({
@@ -89,12 +89,12 @@ t.test('trust-github with file as path', async t => {
   })
 
   await t.rejects(
-    npm.exec('trust', ['github', packageName, '--yes', '--file', '.github/workflows/ci.yml', '--repository', 'owner/repo']),
+    npm.exec('trust', ['gitlab', packageName, '--yes', '--file', '.gitlab/ci.yml', '--project', 'group/repo']),
     { message: /must be just a file not a path/ }
   )
 })
 
-t.test('trust-github without environment', async t => {
+t.test('gitlab without environment', async t => {
   const { npm } = await loadMockNpm(t, {
     prefixDir: {
       'package.json': JSON.stringify({
@@ -124,30 +124,30 @@ t.test('trust-github without environment', async t => {
 
   registry.trustCreate({ packageName })
 
-  await npm.exec('trust', ['github', packageName, '--yes', '--file', 'workflow.yml', '--repository', 'owner/repo'])
+  await npm.exec('trust', ['gitlab', packageName, '--yes', '--file', '.gitlab-ci.yml', '--project', 'group/repo'])
 })
 
 t.test('bodyToOptions with all fields', t => {
-  const TrustGitHub = require('../../../lib/subcommands/trust-github.js')
+  const TrustGitLab = require('../../../../lib/commands/trust/trust-gitlab.js')
 
   const body = {
     id: 'test-id',
-    type: 'github',
+    type: 'gitlab',
     claims: {
-      repository: 'owner/repo',
-      workflow_ref: {
-        file: 'test.yml',
+      project_path: 'group/repo',
+      ci_config_ref_uri: {
+        file: '.gitlab-ci.yml',
       },
     },
     environment: 'prod',
   }
 
-  const options = TrustGitHub.bodyToOptions(body)
+  const options = TrustGitLab.bodyToOptions(body)
 
   t.equal(options.id, 'test-id', 'id should be set')
-  t.equal(options.type, 'github', 'type should be set')
-  t.equal(options.file, 'test.yml', 'file should be set')
-  t.equal(options.repository, 'owner/repo', 'repository should be set')
+  t.equal(options.type, 'gitlab', 'type should be set')
+  t.equal(options.file, '.gitlab-ci.yml', 'file should be set')
+  t.equal(options.project, 'group/repo', 'project should be set')
   t.equal(options.environment, 'prod', 'environment should be set')
   t.end()
 })
